requirements.txt
discord.py==2.3.2
openai==1.16.1
python-dotenv==1.0.1